signal_handler: Implementing specific messages per context
As per bug #23320, regarding the signal we receive, it could be interesting providing a specific message.

If we get a SI_USER under SIGTERM or SIGINT, then we almost print the actual message except
we don't print the PID which have no meaning when set to 0.

In all other cases, we do print the full structure to help debuggers understanding the received signal.

This patch is offering a better way to get specific messages based on various signals or si_code values.
It could be expanded later based on debugging feedbacks.

Fixes: #23320